refactor: playWithDoki to UserItem component
