+++ src/Components/dokiPack/UserCarrots.js
@@ -1,10 +1,5 @@
-import React, {useState} from "react";
-import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-} from 'react-native';
+import React, { useState } from "react";
+import { StyleSheet,Text, View, TouchableOpacity } from 'react-native';
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
 import { Popable, usePopable } from 'react-native-popable';
 import { useQueryClient } from 'react-query';
