+++ src/Components/dokiPack/UserItem.js
@@ -1,7 +1,11 @@
-import { StyleSheet, View, Image, Text, TouchableOpacity } from 'react-native';
-import React from 'react';
-import { Card, Title, Paragraph } from 'react-native-paper';
-import { StyledItemCard, StyleItemImage } from '../styles';
+import React, { useState } from "react";
+import { StyleSheet, View, Text, TouchableOpacity } from 'react-native';
+import { StyleItemImage } from '../styles';
+import { Popable, usePopable } from 'react-native-popable';
+import { useQueryClient } from 'react-query';
+import { useUpdateUserDoki } from '../../hooks/useUpdateUserDoki';
+import { useUpdateUser } from '../../hooks/useUpdateUser';
+import { createTriggerNotification } from "../../helpers/createTriggerNotification";
 
 const imageNames = {
   'video game': require('../../../assets/items/videogame.png'),
@@ -12,19 +16,51 @@ const imageNames = {
   paintbrush: require('../../../assets/items/paintbrush.png'),
 };
 
-const UserItem = (props) => {
-  // console.log('PROPS INSIDE USER ITEM', props);
+const UserItem = ({name, quantity, curMoodLvl}) => {
+  const { ref, hide, show } = usePopable();
+  const userDokiMutation = useUpdateUserDoki();
+  const userMutation = useUpdateUser();
+  const queryClient = useQueryClient();
+  const [msgContent, setMsgContent] = useState(null);
 
   return (
-    <TouchableOpacity onPress={props.handlePlay}>
+    <TouchableOpacity onPress={playWithDoki}>
       <View style={styles.box}>
-        <StyleItemImage source={imageNames[props.name]} />
+        <StyleItemImage source={imageNames[name]} />
         <View style={styles.quantity}>
-          <Text style={styles.text}>{props.quantity}</Text>
+          <Text style={styles.text}>{quantity}</Text>
         </View>
       </View>
+      <Popable
+        ref={ref}
+        content={msgContent}
+        style={styles.popable}
+        animationType="spring"
+      ></Popable>
     </TouchableOpacity>
   );
+
+  function playWithDoki () {
+    if (curMoodLvl >= 100) {
+      setMsgContent("I'M ALL PLAYED OUT!");
+    } else {
+      const newMoodLevel = curMoodLvl + 5;
+      const userDokiUpdate = {
+        lastPlayedAt: new Date(),
+        lastPlayedMoodLevel:
+          curMoodLvl + (newMoodLevel > 100 ? 100 - curMoodLvl : 5), // Mood Increase Rate
+      };
+      userDokiMutation.mutate(userDokiUpdate, {
+        onSuccess: () => {
+          queryClient.invalidateQueries(['userDoki'])
+        },
+      });
+      setMsgContent('THIS IS SO MUCH FUN!');
+      show();
+      setTimeout(() => hide(), 1000);
+      createTriggerNotification('play');
+    }
+  };
 };
 
 export default UserItem;
@@ -61,5 +97,10 @@ const styles = StyleSheet.create({
     borderRadius: 100,
     paddingTop: 2,
     marginLeft: 'auto',
+  },
+  popable: {
+    alignSelf: 'center',
+    marginTop: 330,
+    width: 200,
   },
 });
